lib/os/heap: Correct aligned_alloc sizing for small heaps
The code that made aligned_alloc work with the 4-byte heap headers was
correctly requesting a block of the correct size, and correctly
aligning the output buffer within that memory, but it was using the
UNALIGNED chunk size for the buffer as the final size of the block
with splitting off the unused suffix.  So the final chunk in the
buffer was being incorrectly returned to the heap and reused, leading
to overlap.

Compute the chunk size of the output buffer based on the
already-aligned output pointer instead.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>