soc/intel_adsp: Unmask IDC interrupts unconditionally in arch_start_cpu()
There is some existing code for the older hardware to deal with a race
between the IDC protocol used by the boot ROM and the one used by the
OS.  It will mask the IDC interrupt immediately after sending the ROM
message, and then unmask it on the CPU startup path.  This seems to
work, but not in the SOF app itself.

It seems like SOF still has some vestigial code doing this itself,
presumably at a different spot or with a different delay which is
defeating the Zephyr-side unmasking.  Unmask immediately before
sending if we're building for SOF as a workaround.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>