soc/intel_adsp: Unify ROM IDC delay handling
On older hardware, the core startup message and scheduler IPI use the
same underlying IDC interurpt.  There is a mask step in
arch_start_cpu() to avoid confusing the two.  And there's a delay to
make sure the ROM has time to see it before disabling the interrupt.

Effectively this same delay was appearing on the core startup path,
where we needed to wait for the interrupt to be masked by the main
core before unmasking it on ours (otherwise it would be left
unmasked!).

Except the connection between the two had gotten lost
archaeologically, so this one was labelled "voodoo".  It was
verifiably needed, I just didn't know why.  Well, now I know why.

Clean up the comment.  Use the same delay on both sides.  And flag
with the correct platform macro (!cavs25 and not just cavs15).

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>