soc/intel_adsp: Remove SOF interrupt workaround
Further investigation by Kai Vehmanen has shown that this isn't the
real issue.  What's happening is that on cAVS 1.5 devices, secondary
core power is managed by the host.  And there is a delay before the
ROM is available (unclear whether this is a race vs. hardware, the ROM
firmware, or the kernel driver, or potentially some combination);
interrupts latched earlier than that seem to be cleared by CPU
initialization.

Just a delay at the start of arch_start_cpu() is all that's needed.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>