drivers: Add "inotice" abstraction, implement via cavs-ipc
There's desire to have a generic abstraction to do what the SOF IPC
layer does (currently implemented with its own driver on the same
underlying device), so that future SOF platforms have a clean
interface to target.

This adds "inotice" (Interrupt Notice), which simply splits out the
most obvious parts of IPC (sending a synchronous interrupt with two
words of data) as a subsystem API, and forwards to the existing
cavs_ipc driver via callbacks in the device model API struct.

The full cavs_ipc device API remains available for code that wants to
exploit the more oddball features like return "DONE" interrupts and
asynchronous completion.  The same device pointer is used for both, of
course.

But SOF and all reasonable apps should hopefully be happy with
inotice.

Signed-off-by: Andy Ross <andrew.j.ross@intel.com>