kernel: Replace k_sem with k_zync wrapper
Replace the original semaphore code with a thin wrapper over k_zync.
As for mutex, the default configuration is now leaner, without support
for the "max value" feature of the original code.  That can be
configured in via CONFIG_ZYNC_ATOM_VAL_MAX if desired for backwards
compatibility, but the default is now closer to an industry-standard
interface (and much faster).  The only significant use of the value
clamping was to use a semaphore to implement a lock, and that is
needless now that k_mutex is also a lean zync.

Signed-off-by: Andy Ross <andyross@google.com>