kernel: Add k_poll support for zync
Add support for polling on zync objects via the standard API.

FIXME: this patch is currently a bisection problem as poll is broken
between the k_sem wrapper patch before this and this patch.  It needs
to be split to separate "zync poll support" from "zync poll sem
support" so they can go in in the right order.

Signed-off-by: Andy Ross <andyross@google.com>