tests/kernel/workq: Fix racy test on semaphore
The sync_sem has a clamp of 1, and there are two k_sem_give() calls
being executed from higher priority work items that will saturate.
Yet the code was trying to guarantee that two k_sem_take() calls would
return zero.  That's not possible per the API spec unless the test is
trying to whitebox exactly where scheduling points might happen,
something that we don't document.

Just remove.  This doesn't look correct, and the change to the way
rescheduling works in zync is tripping over this.

Signed-off-by: Andy Ross <andyross@google.com>