include/toolchain: Clarify docs for STRUCT_SECTION_ITERABLE
The need for linker script integration was always a little bit of a
kluge, but now we've grown a wart where there are TWO different linker
script generation schemes and BOTH need to be told about the sections.

Make it super duper clear in the docs.  And include a whine for
posterity that there's no need for this and that you can see these
section name schemes in the object files, we should be able to do this
automatically.

Signed-off-by: Andy Ross <andyross@google.com>