samples/userspace/shared_mem: Correct scheduling behavior
This sample spawns three threads, all of which spin in loops doing
work without waiting on anything but a single semaphore used as a
mutex.

But without timeslicing enabled, if a thread doesn't block on the
semaphore, it will never leave the run queue and thus never release
the CPU!  So with fewer than three CPUs, this test was theoretically
able to starve a thread and live lock.

As it happened, the implementation of k_sem_take() always acted like a
scheduling point, so the test never got stuck.  But this behavior was
never documented, and newer synchronization mechanisms like zync don't
implement it like that (in fact zync is a 100% kernel free atomic
operation that can't possibly block in the uncontended case).

Signed-off-by: Andy Ross <andyross@google.com>