net/context: Initialize mutex earlier
The k_mutex lock on new network context objects was being initialized
too late, after other initialization code was trying to use it.  Push
it up to the top, right after the memset.  (This is a soft failure
with traditional k_mutex, but when ZYNC_VALIDATE=y it becomes an
assertion).

Signed-off-by: Andy Ross <andyross@google.com>