samples/cmsis_v*/philosophers: Disable SMP
These are cut-and-paste copies of the very old philosophers demo, but
with CMSIS synchronization primitives.  This code has never been made
MP-safe, and the original runs in uniprocessor mode.  This missed that
config, so it's quite unstable on SMP machines.  In CI that's covered
up by our retries, but we shouldn't even be trying.

(Really we should probably merge these with the original, there's no
good reason to keep three copies of the same mostly-identical files in
the tree)

Signed-off-by: Andy Ross <andyross@google.com>