tests/kernel/sched/preempt: ARC can't irq_offload
The ARC architecture has the same issue that ARM32 got called out for
in this test (which I think is fixed?), where irq_offload() doesn't
honor scheduler state on return from the handler.  And as elsewhere,
rescheduling points in the older synchronization primitives was saving
us; zync makes these ops lockless and exposes the bug.

Signed-off-by: Andy Ross <andyross@google.com>