lib/os/fdtable: Correctly initialize POSIX fd's
The file descriptors for stdin/out/err are static in the table, but
the k_mutex object they contain cannot be statically initialized.
This works with k_mutex, but zync is stricter and will deadlock.  Put
a light weight lazy-evaluated call to k_mutex_init() into the existing
descriptor validation hook.

Signed-off-by: Andy Ross <andyross@google.com>