tests/net/socket/udp: Semaphores should be semaphores
This test was using a sys_mutex as a "wakeup source".  One thread
would "lock" it intending to go to sleep, for another to "unlock" it.
That's sort of an abuse of the API intent.  Semaphores are the
appropriate abstraction for that use case.  And with Zync, the attempt
to unlock a mutex you don't own is a fatal assertion.

Signed-off-by: Andy Ross <andyross@google.com>