tests/benchmarks/latency_measure: Needs recursive zyncs
This test tries to bench mark a k_mutex by iteratively locking it 1000
times in a row.  Obviously if you try that on a non-recursive mutex,
you'll deadlock on the second call.

So enable recursion.  But that's only a half solution, because now it
means our benchmark app is benchmarking our mutex primitive in its
slowest configuration!  Needs rework.

Signed-off-by: Andy Ross <andyross@google.com>