tests/sys_mutex: Fix scheduling/priority behavior
This spot would unlock a mutex from a high priority thread (priority
10 in mutex_complex()) that was being waited on by a low priority
thread (12/thread_12()), then expect that the thread would be able to
acquire it before the higher priority thread tried to lock it again.
Older k_mutex code would act as a scheduling point (and on SMP
platforms another CPU could race in to "fix" it), but I don't see why
this ever should have worked.  Put a sleep in to allow the low
priority worker to reach sys_mutex_lock() reliably.

Similarly there was a spot in test_mutex_multithread_competition()
where the code was assuming that the thread created in
k_thread_create() would have a chance to run and pend on the mutex
before the main thread released it to start the test.

Signed-off-by: Andy Ross <andyross@google.com>