arch/xtensa: Fix fatal error mask level
This isn't quite correct.  An NMI is an unmaskable interrupt, so it's
not possible to set the INTLEVEL PS field to XCHAL_NMILEVEL.  The
correct way to express "maximum interrupt level" is XCHAL_EXCM_LEVEL
(the level masked by the EXCM bit).

Actually on supported hardware, this works fine as it clamps to
whatever the real maximum is anyway.  The bigger problem is that not
all cores enable an NMI, and so core-isa.h doesn't always define
XCHAL_NMILEVEL, causing build errors.

Signed-off-by: Andy Ross <andyross@google.com>