arch/arm/cortex_m: Set thumb bit in new thread contexts
Hardware exception handling on ARM Cortex-M ignores the bottom bit of
the instruction pointer popped from the frame.  But the hardware is
thumb-only, and software users (like evolved context switching) need
this set to mark thumb state.  Software jumps to such an address will
trap with a UsageFault.

Set the bit on thread creation.

Signed-off-by: Andy Ross <andyross@google.com>