soc/mediatek/mtk_adsp: Move stacks to SRAM
Add a tiny feature at the top level allowing a new section (".stacks")
for all stack memory (vs. .noinit and/or .user_stacks) such that this
device can place them in SRAM for performance.  Longer term we should
come up with a unified plan for "stack placement" such that it can
coexist with USERSPACE and KERNEL_COHERENCE, which also want control
over stack memory at the linker level.  But this is very simple.

Signed-off-by: Andy Ross <andyross@google.com>